+++ nodes/blynk-iot-in-app.js
@@ -10,35 +10,35 @@ module.exports = (RED) => {
     if (this.blynkClient) {
       this.blynkClient.registerInputNode(this);
 
-      this.blynkClient.on('opened', (n) => { // eslint-disable-line no-shadow
+      this.blynkClient.on('status-connecting', () => { // eslint-disable-line no-shadow
         node.status({
           fill: 'yellow',
           shape: 'dot',
-          text: RED._('blynk-iot-in-app.status.connecting') + n,
+          text: RED._('blynk-iot-in-app.status.connecting'),
         });
       });
-      this.blynkClient.on('connected', () => {
+      this.blynkClient.on('status-connected', () => {
         node.status({
           fill: 'green',
           shape: 'dot',
           text: 'blynk-iot-in-app.status.connected',
         });
       });
-      this.blynkClient.on('error', () => {
+      this.blynkClient.on('status-error', () => {
         node.status({
           fill: 'red',
           shape: 'ring',
           text: 'blynk-iot-in-app.status.error',
         });
       });
-      this.blynkClient.on('closed', () => {
+      this.blynkClient.on('status-disconnnected', () => {
         node.status({
           fill: 'red',
           shape: 'ring',
           text: 'blynk-iot-in-app.status.disconnected',
         });
       });
-      this.blynkClient.on('disabled', () => {
+      this.blynkClient.on('status-disabled', () => {
         node.status({
           fill: 'red',
           shape: 'dot',