+++ nodes/blynk-iot-style-btn.js
@@ -18,35 +18,35 @@ module.exports = (RED) => {
     this.blynkClient = RED.nodes.getNode(this.client);
     if (this.blynkClient) {
       this.blynkClient.registerInputNode(this);
-      this.blynkClient.on('opened', (n) => { // eslint-disable-line no-shadow
+      this.blynkClient.on('status-connecting', () => { // eslint-disable-line no-shadow
         node.status({
           fill: 'yellow',
           shape: 'dot',
-          text: RED._('blynk-iot-style-btn.status.connecting') + n,
+          text: RED._('blynk-iot-style-btn.status.connecting'),
         });
       });
-      this.blynkClient.on('connected', () => {
+      this.blynkClient.on('status-connected', () => {
         node.status({
           fill: 'green',
           shape: 'dot',
           text: node.connected_label,
         });
       });
-      this.blynkClient.on('error', () => {
+      this.blynkClient.on('status-error', () => {
         node.status({
           fill: 'red',
           shape: 'ring',
           text: 'blynk-iot-style-btn.status.error',
         });
       });
-      this.blynkClient.on('closed', () => {
+      this.blynkClient.on('status-disconnnected', () => {
         node.status({
           fill: 'red',
           shape: 'ring',
           text: 'blynk-iot-style-btn.status.disconnected',
         });
       });
-      this.blynkClient.on('disabled', () => {
+      this.blynkClient.on('status-disabled', () => {
         node.status({
           fill: 'red',
           shape: 'dot',
@@ -60,6 +60,7 @@ module.exports = (RED) => {
     this.on('input', function input(msg) {
       // no input operation if client not connected or disabled
       if (!node.blynkClient || !node.blynkClient.logged) {
+        node.log(`## logged is FALSE`);
         return;
       }